+++ src/main/java/com/owlmaddie/chat/PlayerData.java
@@ -9,8 +9,12 @@ package com.owlmaddie.chat;
  */
 public class PlayerData {
     public int friendship;
+    public long lastDamageTime;
+    public long swingStartTime;
 
     public PlayerData() {
         this.friendship = 0;
+        this.lastDamageTime = 0;
+        this.swingStartTime = 0;
     }
 }